Add nil strategy and refactor to include network name in upgrades.
