+++ app/upgrades/mainnet/v1.17.0/rc2/upgrades.go
@@ -1,4 +1,4 @@
-package v1_18_0
+package rc2
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/provenance-io/provenance/app/keepers"
 	"github.com/provenance-io/provenance/app/upgrades"
-	rc1 "github.com/provenance-io/provenance/app/upgrades/v1.18.0/rc1"
+	"github.com/provenance-io/provenance/app/upgrades/mainnet/v1.17.0/common"
 )
 
 func UpgradeStrategy(ctx sdk.Context, app upgrades.AppUpgrader, vm module.VersionMap) (module.VersionMap, error) {
@@ -20,9 +20,10 @@ func UpgradeStrategy(ctx sdk.Context, app upgrades.AppUpgrader, vm module.Versio
 		return nil, err
 	}
 
-	return newVM, nil
+	return newVM, err
 }
 
-func PerformUpgrade(ctx sdk.Context, k *keepers.AppKeepers) (err error) {
-	return rc1.PerformUpgrade(ctx, k)
+func PerformUpgrade(ctx sdk.Context, k *keepers.AppKeepers) error {
+	common.UpdateIbcMarkerDenomMetadata(ctx, k)
+	return nil
 }
